+++ src/cider/nrepl.clj
@@ -213,18 +213,27 @@ Depending on the type of the return value of the evaluation this middleware may 
    "doc-first-sentence-fragments" (str "May be absent. Represents the first sentence of a Java doc comment. " fragments-desc)
    "doc-block-tags-fragments"     (str "May be absent. Represent the 'param', 'returns' and 'throws' sections a Java doc comment. " fragments-desc)})
 
+(def info-params
+  {"sym"     "The symbol to lookup"
+   "ns"      "The current namespace"
+   "context" "A Compliment completion context, just like the ones already passed for the \"complete\" op,
+with the difference that the symbol at point should be entirely replaced by \"__prefix__\".
+
+For Java interop queries, it helps inferring the precise type of the object the `:sym` or `:member` refers to,
+making the results more accurate (and less numerous)."
+   "class"  "A Java class. If `:ns` is passed, it will be used for fully-qualifiying the class, if necessary."
+   "member" "A Java class member."})
+
 (def-wrapper wrap-info cider.nrepl.middleware.info/handle-info
   (cljs/requires-piggieback
    {:requires #{#'session}
     :handles {"info"
               {:doc "Return a map of information about the specified symbol."
-               :requires {"sym" "The symbol to lookup"
-                          "ns" "The current namespace"}
+               :optional info-params
                :returns (merge {"status" "done"} fragments-doc)}
               "eldoc"
               {:doc "Return a map of information about the specified symbol."
-               :requires {"sym" "The symbol to lookup"
-                          "ns" "The current namespace"}
+               :optional info-params
                :returns (merge {"status" "done"} fragments-doc)}
               "eldoc-datomic-query"
               {:doc "Return a map containing the inputs of the datomic query."